requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0

# TensorBoard (for training monitoring)
tensorboard>=2.13.0

# Kaggle Hub (for dataset download)
kagglehub>=0.2.0

# Optional: For Jupyter notebooks if you create examples
# jupyter>=1.0.0
# ipywidgets>=8.0.0
